fix currency symbols in other components, move to IOUUtils
